requirements.txt
protobuf~=5.29.3
python-dotenv~=1.0.1
ragas~=0.2.13
langchain-google-genai~=2.0.9
playwright~=1.50.0
behave~=1.2.6